requirements.txt
asgiref==3.4.1
Django==3.2.5
djangorestframework~=3.12.4
pytz==2021.1
requests~=2.26.0
sqlparse==0.4.1
